Replace doubled quantifier by right-to-left matching
The duplicate quantifiers will still cause execution time of the regex
to grow exponentially, albeit in a much more manageable way.
By starting matching from the right, a valid match is returned as soon
as the first 's3' is encountered, which eliminates this as well and cuts
down execution time further.