Fix wrong assertion in `CopyObject_Test7`
The Minio-server is returning an error for a specific object (`"MinIO API
responded with message=At least one of the pre-conditions you specified did not hold for object: \"/minio-dotnet-
example-3uvmhkjpxp0ndmg/minio-dotnet-example-yhrrlm29wq\""`). Therefore, the error message should not be exactly the expected string, but rather it should **start** with the expected string.