Check if bucket exists before attempting to delete it
If you attempt to delete a bucket if it does not exist, the API returns a `404` for the not existent bucket, which is correct. So before the attempt to remove the bucket, there should be a check if it even exists.

This test-bug was introduced in commit 9772a373b92d81808118edbdbd653d301f4502ac, which also disabled the error handling for all requests - which is the reason this bug was not discovered so far.